requirements.txt
aiohttp
async-timeout
attrs
chardet
discord.py
idna
multidict
psutil
python-dotenv
websockets
yarl